Moved avatar-img and custom-btn to shared
