--- src/app/components/answer-label/answer-label.component.css
@@ -1,23 +1,0 @@
-:host {
-  display: contents;
-}
-
-.answer {
-  font-size: 1.75rem;
-  background-color: var(--secondary-color-dark);
-  border-radius: var(--radius-btn);
-  border: solid 4px currentColor;
-  font-weight: bold;
-  height: 5rem;
-  min-width: 20.4375 rem;
-  display: grid;
-  place-content: center;
-
-  &:hover {
-    border-color: var(--accent-color);
-  }
-}
-
-.active-answer {
-  border-color: var(--accent-color);
-}
