--- src/app/components/answer-label/answer-label.component.html
@@ -1,6 +1,0 @@
-<!-- //TODO data pendiente: ImageParameters  -->
-<div
-  class="answer text-center text-capitalize shadow-md"
->
-  {{ answer() || 'empty' }}
-</div>
