+++ src/app/pages/home/home.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { LogoTitleComponent } from '../../shared/components/logo-title/logo-title.component';
-import { AvatarImageComponent } from '../../components/avatar-img/avatar-image.component';
-import { CustomButtonComponent } from '../../components/custom-btn/custom-button.component';
+import { AvatarImageComponent } from '../../shared/components/avatar-img/avatar-image.component';
+import { CustomButtonComponent } from '../../shared/components/custom-btn/custom-button.component';
 
 export interface Player{
         id?: number,
