+++ src/app/shared/components/avatar-img/avatar-image.component.spec.ts
