--- src/app/shared/components/btn-basic/btn-basic.component.css
@@ -1,3 +1,0 @@
-:host {
-  display: block;
-}
